# openapi-spec.yaml
openapi: 3.0.3
info:
  title: DocPort.io API
  description: DocPort.io API documentation
  version: 0.0.1
servers:
  - url: 'http://localhost:8080'
paths:
...
  /api/v1/users:
    post:
      operationId: createUser
      summary: Create a new user
      tags:
        - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        409:
          $ref: '#/components/responses/Conflict'
        500:
          $ref: '#/components/responses/InternalServerError'
  /api/v1/users/{userId}:
    get:
      operationId: getUserById
      summary: Get a user by ID
      tags:
        - users
      parameters:
        - $ref: '#/components/parameters/PathUserId'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
  /api/v1/users/me:
    get:
      operationId: getAuthenticatedUser
      summary: Get the authenticated user
      tags:
        - users
      security:
        - OpenIdConnect: []
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    OpenIdConnect:
      type: openIdConnect
      openIdConnectUrl: 'https://keycloak.docport.io/realms/docport-dev/.well-known/openid-configuration'
  parameters:
...
    PathUserId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
        format: int64
  responses:
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...
    UserResponse:
      type: object
      required:
        - id
        - createdAt
        - updatedAt
        - name
        - email
        - emailVerified
      properties:
        id:
          type: integer
          format: int64
          example: 1
        createdAt:
          type: string
          format: date-time
          example: '2026-01-01T00:00:00.000Z'
        updatedAt:
          type: string
          format: date-time
          example: '2026-01-01T00:00:00.000Z'
        name:
          type: string
          example: 'John Doe'
        email:
          type: string
          format: email
          example: john.doe@example.com
        emailVerified:
          type: boolean
          example: true
    CreateUserRequest:
      type: object
      required:
        - name
        - email
        - emailVerified
      properties:
        name:
          type: string
          example: 'John Doe'
        email:
          type: string
          format: email
          example: john.doe@example.com
        emailVerified:
          type: boolean
          example: true
...